app_prototype-master/requirements.txt
nicegui==2.22.2
argon2-cffi==25.1.0
requests==2.32.4
pandas==2.3.1
reportlab==4.4.4
rapidfuzz==3.14.1